requirements.txt
python-dotenv
uvicorn
fastapi
deepl
transformers
torch
psutil
gdown